Fix integer overflow in PhysicalState.ScaleChangeToHealth
amount * range could exceed Int32 for large (firepower-boosted) changes, wrapping negative and flipping the relaxation state's sign, which showed high-HP buildings with a spurious cryo overlay. Use 64-bit math for the intermediate product.

+++ OpenRA.Mods.Common/Traits/PhysicalState.cs
@@ -275,7 +275,10 @@ namespace OpenRA.Mods.Common.Traits
 		public int ScaleChangeToHealth(int amount)
 		{
 			if (health == null || health.MaxHP == 0 || !Info.RelativeToHealth) return amount;
-			return amount * range / health.MaxHP;
+
+			// Use 64-bit math: amount * range can exceed int range for large changes
+			// (e.g. firepower-boosted hits), which would wrap negative and flip the state's sign.
+			return (int)((long)amount * range / health.MaxHP);
 		}
 	}
 }
